module1/requirements.txt
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
astroid==2.5.1
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gpg==1.10.0
httplib2==0.9.2
idna==2.6
isort==5.8.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
meld==3.18.0
mysqlclient==2.0.3
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==2.0.1
pylint==2.7.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-magic==0.4.16
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scour==0.36
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.4.1
system-service==0.3
systemd-python==234
TLPUI==1.3.1.post3
toml==0.10.2
typed-ast==1.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wrapt==1.12.1
xkit==0.0.0
zope.interface==4.3.2
